Fixed query params ordering in AWS signature
Also moved tests so that they are executed on Native and JS
