+++ modules/aws-http4s/test/src/smithy4s/aws/AwsCredentialsFileTest.scala
@@ -17,8 +17,6 @@
 package smithy4s.aws
 
 import weaver._
-import fs2.io.file.Files
-import cats.effect.IO
 import smithy4s.aws.kernel.AwsCredentials
 
 object AwsCredentialsFileTest extends FunSuite {
@@ -129,43 +127,4 @@ object AwsCredentialsFileTest extends FunSuite {
   }
 
   private def asLines(s: String) = s.split("\\n").toList
-}
-
-object AwsCredentialsFileReadTest extends SimpleIOSuite {
-  test("read credentials from a file") {
-    Files[IO].tempFile.use { path =>
-      val content = """|[default]
-                       |aws_secret_access_key = def_sec
-                       |aws_access_key_id     = def_key
-                       |aws_session_token     = def_token
-                       |[profile p1]
-                       |aws_secret_access_key = sec
-                       |aws_access_key_id     = key
-                       |aws_session_token     = token
-                  """.stripMargin
-
-      val writeFile = fs2.Stream
-        .emit(content)
-        .through(fs2.text.utf8.encode)
-        .through(Files[IO].writeAll(path))
-        .compile
-        .drain
-      val readDefault = AwsCredentialsFile.fromDisk(path, None)
-      val readP1 = AwsCredentialsFile.fromDisk(path, Some("p1"))
-
-      for {
-        _ <- writeFile
-
-        default <- readDefault
-        p1 <- readP1
-      } yield {
-        val expectedP1 =
-          AwsCredentials.Default("key", "sec", Some("token"))
-        val expectedDefault =
-          AwsCredentials.Default("def_key", "def_sec", Some("def_token"))
-        expect.same(expectedDefault, default) &&
-        expect.same(expectedP1, p1)
-      }
-    }
-  }
 }
