+++ modules/aws-http4s/test/src/smithy4s/aws/AwsErrorTypeDecoderTest.scala
